extracted Expectations to a CoreExpectation class
